Parameter targetEncoding will select the encoding of the generated sourcefiles
When the machine on which the build is started has odd locale settings,
the default encoding of the VM will be inherited and could arise
encoding issues. The new TargetEncoding option can be set in the 
pom-configuration-section, for example to UTF-8.
Secondly the comments of enum values are handled properly.